move all auth stuff into user module
